+++ src/Blitz.php
@@ -283,7 +283,8 @@ class Blitz extends Plugin
         Event::on(Elements::class, Elements::EVENT_BEFORE_DELETE_ELEMENT,
             function(DeleteElementEvent $event) {
                 if ($event->hardDelete && $event->element !== null) {
-                    $this->refreshCache->addCacheIds($event->element);
+                    $cacheIds = $this->refreshCache->getElementCacheIds([$event->element->getId()]);
+                    $this->refreshCache->addCacheIds($cacheIds);
                 }
             }
         );